# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AI-HOS Hospital Operating System API
  description: |
    REST API for AI-HOS — a multi-tenant, AI-powered Hospital Operating System.

    ## Authentication
    All protected endpoints require a valid session cookie from NextAuth.
    The session includes `tenantId`, `role`, and user identity claims.

    ## Multi-Tenancy
    All data is scoped by `tenant_id`. The tenant is injected from the
    authenticated session — clients cannot specify a different tenant
    (except SUPER_ADMIN who can switch branches).

    ## Rate Limiting
    - OTP endpoints: 3 per phone per 15 minutes, 10 per IP per 15 minutes
    - WhatsApp webhook: 30 messages per user per 5 minutes
    - Login attempts: 5 per 15 minutes
  version: 1.0.0
  contact:
    name: AI-HOS Support
servers:
  - url: /api
    description: API routes
...
paths:
...
  /whatsapp/webhook:
    get:
      tags: [WhatsApp]
      summary: Meta webhook verification
      security: []
      parameters:
        - name: hub.mode
          in: query
          schema: { type: string }
        - name: hub.verify_token
          in: query
          schema: { type: string }
        - name: hub.challenge
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Challenge response
        '403':
          description: Verification failed
    post:
      tags: [WhatsApp]
      summary: Receive WhatsApp messages
      security: []
      description: |
        Receives messages from Meta WhatsApp Cloud API.
        Always returns 200 to prevent Meta retry storms.
      responses:
        '200':
          description: Always returns ok
...